First commit of fix for LogarithmicSI column
The WSClean 2.4 source list contains a LogarithmicSI column which is not compatible with current lofar software makesourcedb. This edit to the filter_model script adds the ability to remove this column from the sky model while also accounting for multi component spectral indices. Tested manually but not yet in pipeline environment