Major Change - Delete Column separate script
The function to delete the LogarithmicSI column is now a standalone
script. This is run before the filter_model steps in the parset. These
steps can simply be removed when a new version of makesourcedb is
available in the future.

Original steps are kept and commented out.

filter_model.py has been returned to its original state.